A Walk Through the Past
Final Assignment for Internet Mapping 3rd Course:

Take a walk through a land that has been occupied for over 10,000 years. This map shows popular trails throughout Simcoe County, Ontario, Canada and their relation to the occupational and land-use history of their surroundings.

Detailing a collection of more than 100 archaeological sites documented in the past 125 years, this map serves to edify the presence and perseverance of the Indigenous groups in the face of an ever-shifting landscape. Shorelines of the variations of Glacial Lake Algonquin and Lake Huron are shown at 3 different iterations reflecting the availability of accessible land. All this is shown on a modern topographical map to show the relation of the ancient peoples to modern settlements.